[r=allenap][bug=1030881][author=rvb] Fix dhcp config template.  Gavin is right, the pool section is used to allocate a sub range for a particular class of client.
I've tested that the config file is syntactically valid by generating a config file using reasonable values for all the parameters and testing that a dhcp server starts correctly using that file. Something I learned in the process is that, by default, the server does not listen on any interface. The interface needs to be added to /etc/default/isc-dhcp-server.  That's work for another branch.