[r=rvb][bug=][author=jtv] Add --dhcp_interface option to “maas configure_master_dhcp,” to select the network interface that dhcpd should service.
No real code is needed, since this stuff is all data-driven. The option just magically appears. The one complication is that being a text field in a Django app, dhcp_interface uses the empty string for null — it's allowed to be blank, but it has a not-null constraint on it so we mustn't leave it None.