[r=rvb][bug=][author=jtv] Support function: write custom config section.  Maintains a custom MAAS section in config file text.
Needed for updating local extensions to AppArmor profile for isc-dhcp-server, so we can run it with our own pidfile and leases file.  The new function appends, or replaces, a clearly marked custom MAAS section in a configuration-file text.  It seems like a classic problem but discussion with a great number of people turned up no small, reusable alternative.  Nowadays we prefer runparts directories, of course, but in this case there is no such facility.  Discussion with Andres Rodriguez, Scott Moser, and others led us to the conclusion that this was our best choice for the given problem.

The function tries to be reasonably robust, so it will try to do something sensible rather than fail when it finds something unexpected.  But it can't be too clever or it may confuse and upset humans whose files it messes with.