[r=julian-edwards][bug=][author=jtv] New maas-provision command: customize-config.  Creates/manages a custom section in a config file.
The use-case at hand is AppArmor config for writing our own dhcpd.  We went through several rounds of progressive insight on this.  The latest insight is that we shouldn't be writing our own config to the existing "local customizations" file; we should merely add an #include referring to a file of our own.  The real work is for packaging to do.  And we must be able to disable that on uninstall, or purging (not merely uninstalling) maas-dhcp would break dhcpd more or less permanently.

You run the custom command by giving it the name of a config file you wish to customize, and feeding it the text for your custom section on standard input.  It will print an updated version of the file to standard output, containing your custom section.

All text manipulation here is done in unicode.  A single encoding is assumed (and it defaults to UTF-8).  One test runs the script in a sub-process, mainly to ensure there are no mistakes in the assumption that stdin/stdout are raw bytes which might lead to failure while dedocing/encoding them.