[r=jameinel][bug=1049407][author=jtv] Document DEFAULT_MAAS_URL as facing the MAAS (rather than the user's browser).
The only place we found where we use DEFAULT_MAAS_URL as an “outward”-facing address is where the packaging tells the user where the UI is.  After that, it's purely for helping the various systems in the MAAS find the central server.  This led to a bug report that turned out to be due to the system not being configured properly.  There were no changes to make, but I found some comments that could be clearer.