[r=rvb][bug=][author=allenap] Update the API description document to better reflect the pertinent handlers and actions.
Removes the ?op=... parameter being added to many URIs, because it's not always relevant; sometimes the op must be passed in the (multipart) body content.

Add a `rest` field to an action description. This indicates if the action is ReSTful or not. All custom operations are defined as *not* ReSTful.

Only declare the URI template and its parameters on the handler. Previously they were repeated for every action.