[r=racb][bug=][author=allenap] Get txlongpoll and dns development services working again.
This is unanticipated fallout from recent changes to buildout.cfg to remove its responsibility for production dependencies. This fix ensures that maastesting is only tested with the common test eggs, and also works around a bug in buildout where it generated a script that attempted to update sys.path with an invalid list literal.