[r=andreserl][bug=1061977][author=smoser] on intel, do not send 'console=ttyS0' (or console=tty1)
as described in LP: #1061977, passing console=ttyS0 to a system that does not
have serial hardware can result in writes to /dev/console failing.  During
enlistment and commissioning, that can also cause those stages to fail as
the enlistment/commissioning process tries to write to stdout, which is attached
to /dev/console by upstart.

This change gives evidence to the requirement for LP: #1044503, as the
only way a user can now add 'console=ttyS0' is by editing files in
/usr/share/pyshared/provisioningserver/pxe which will be overwritten on
package upgrade.