[r=rvb][bug=][author=jameinel] Prefetch the mac_addresses for nodes that get displayed on the nodes/ page and the tags/ page.
The URLs we generate are linked against the mac address, so rather than making N queries, do one large prefetch request. Note that we have to change the syntax in the template, because prefetch will only affect an 'all()' query. However, macaddress_set.reverse can easily be mapped into macaddress_set.all reversed. (Does it even matter if we reverse the mac_addresses, are they stably sorted in the DB?)