[r=gz][bug=][author=jameinel] Improve the performance of nodegroups/$UUID/?op=list-nodes
Tell Django that we only care about the system_id column (which is true). This cuts out a full second of clock time from the api. Best guess is that it is because we are avoiding reading the hardware_details content. It doesn't hurt to filter out the other stuff as well, though.

Also, change the APIs to be accessible to admins (superusers). We allow that for other apis, and it makes it much easier to test their performance, because you can poke at it from maascli instead of having to trigger a worker request.