[r=allenap][bug=1066938][author=rvb] This branch adds the inclusion of the default 'controls' statement so that the init scripts can control the bind server using the default RNDC key from localhost.
= Notes =

It turns out that if not controls statement is provided, "inet 127.0.0.1 port 953 allow { localhost; };" is included silently and this is used by the init scripts to control the bind server.  Since MAAS adds a 'controls' statement to control the bind server, we also need to explicitly include the default 'controls' statement.

I've created a package locally and tested this fix:

Without the default 'controls' statement:

$ sudo /etc/init.d/bind9 restart
 * Stopping domain name service... bind9                                                                                                                       rndc: connect failed: 127.0.0.1#953: connection refused
waiting for pid 14057 to die                [ OK ]
 * Starting domain name service... bind9    [ OK ]

With the default 'controls' statement:

sudo /etc/init.d/bind9 restart
 * Stopping domain name service... bind9                                                                                                                       waiting for pid 13819 to die               [ OK ]
 * Starting domain name service... bind9   [ OK ]