[r=rvb][bug=1116700][author=jtv] Generate CNAME for a node's oldest MACAddress that has a lease.  (We used to generate a CNAME for a node's oldest MACAddress *if it had* a lease).
We don't know why the code was originally written this way, although it was deliberate at the time (originally proposed at https://code.launchpad.net/~rvb/maas/maas-dns2/+merge/114707 ).  Our best recollection now is that it was done that way to mimick how things worked with Cobbler, but it led to problems when network interfaces weren't in the expected order.