[r=rvb][bug=][author=jtv] API: support filtering the networks list by node(s).
Discussed with Julian.  We're supposed to do this set-based as a general API style, but what happens with multiple nodes: do we take the union or the intersection of their networks?  We chose to take the intersection, in other words, return only networks that are attached to all of the given nodes.  This is probably more useful in practice.