[r=rvb][bug=][author=jtv] Base architecture choices for a node on the list of usable architectures (i.e. ones for which boot images are available).  The full choices list (which adds presentable labels) will have to be built by combining this information with what we get from the clusters; for now we can just use the architecture names as labels.
A pair of new test helpers patches the list of usable architectures, so that tests which exercise the node form can easily pass an architecture that form validation will recognise — without having to patch in a fixed enum.  A fixed enum would risk hiding accidental reliance on the old values, failure to cope with a lack of boot images, and so on.