[r=jtv][bug=1305282][author=smoser] enlist_userdata: fix issue if dig fails to reach a dns server
dig exits non-zero and writes a failure message to stdout if it fails to
reach dnsserver. If that occurred, we were assigning that output to
the host variable, and then trying to register the node with that host.

The fix here is to only assign host if the dig command was successful.