[r=julian-edwards][bug=1340208][author=jtv] Fix "NodeGroupInterface has no nodegroup" crash when creating a cluster with unnamed interfaces.
The NodeGroupDefineForm crashes if you make it create not just a cluster but a cluster interface as well, and don't specify a name for the cluster interface.  It's supposed to use the network interface name as (the basis for) the cluster interface name.  The problem was that the NodeGroupDefineForm creates a NodeGroupInterfaceForm with its ‘nodegroup’ field not yet referring to a NodeGroup instance in the database.  Due to the ORM stub, the reference does not evaluate as None — the attempt to dereference it just fails.  The validation code for the cluster interface name refers to the cluster.