[r=jtv,blake-rouse][bug=][author=allenap] New fixture AlternativeClusterRPCFixture that enables end-to-end testing against a stub cluster RPC implementation.
The existing ClusterRPCFixture wires clients up to the "real" cluster RPC implementation, which was a bit short-sighted of me; it is increasingly dependent on external data, which will entail ever more elaborate test fixtures. It's better to work against an accurately stubbed RPC implementation and lean on CI/QA to exercise end-to-end.