[r=rvb][bug=1355014][author=jtv] Replace compose_URL_on_IP with new compose_URL, which accepts an IP address or hostname.
When there is no internet connection, guess_server_address() returns a hostname rather than an IP address, and the config during import took this and tried to construct a URL out of it on the assumption that it was an IP address.  Accepting both fixes an error when running tests without internet connection.