[r=julian-edwards][bug=][author=allenap] The asynchronous decorator can now specify a time-out.
If `timeout` has been specified, the EventualResult returned from a call
from outside of the reactor thread will be waited on for up to `timeout`
seconds. This means that callers don't need to remember to wait. If
`timeout` is `None` then it will wait indefinitely, which can be useful
where the function itself handles time-outs, or where the called
function doesn't actually defer work but just needs to run in the
reactor thread.