[r=gmb][bug=][author=jtv] When allocating static IP addresses for a MAC, do so on each of the cluster interfaces it's attached to.
Typically this will be either just an IPv4 one, or an IPv4 one plus an IPv6 one.  (For now I'm not quite confident enough to rule out multiple IPv6 in the future.)  The code is written not to know about IPv4 and IPv6; all it knows is that the MACAddress can have multiple cluster interfaces and that you want a static IP address on each.  Unless you specify the IP address you want; in that case the address identifies the one cluster interface in whose network that address lies.  The code is written as a series of operations on the set of relevant cluster interfaces, which makes it easier to generalise away the difference between these two kinds of usage as well: allocate one address, or allocate on all cluster interfaces.

Most of the complexity comes from the generalisation of the original method's corner-case behaviour.  The old code raised an error if there was a clash with an IP address that the MAC already had.  In the new situation it's reasonable for a MAC to get, say, a sticky IPv6 address but an automatic IPv4 address: we can't treat that as a clash.  So I generalised the notion of a clash to: "all of the addresses I want have clashes."  If only one of the addresses clashes, the method will just allocate the other.