[r=blake-rouse][bug=][author=allenap] Don't wait for a cluster RPC connection by default.
For now it's safer to immediately return when there's no RPC connection available to a specified cluster. Time-outs can block the web application for long periods of unproductive time. We will be better served by making callers better able to deal with an unstable environment. The option of waiting remains though, for code that really needs it.