[r=allenap][bug=][author=jtv] Hard-code paths to DHCPv6 files, instead of taking them from Celery config.
At the point where these paths are needed, the celery config does not seem to contain them!  I only see a very minimal celery config, not all the settings we define in celeryconfig_common.py.  It's not quite worth digging up, because we're in the process of removing this configuration anyway.