[r=jtv][bug=1354014][author=blake-rouse] Replaced TransactionWrapper with ConnectionWrapper.
TransactionWrapper wrapped a LargeFileObject in its own transaction, when reading data into StreamingHttpResponse. This works in theory but causes the main connection for django to be in a transaction, and when a new request is received the Transaction middleware will fail to initialize do to the connection already being in a transaction.

ConnectionWrapper wraps the LargeFileObject in its own connection to the database. This allows the large object to use its own transaction, and not cause any issues with incoming requests to django.