[r=blake-rouse][bug=][author=rvb] Refactor the usage of states 'BROKEN' and 'FAILED_DEPLOYMENT'.
According to the new node lifecycle, 'BROKEN' is the state a node gets into when a user deliberately marks a node as broken to indicate that an admin needs to do something with this node.

'FAILED_DEPLOYMENT' is what happens when a the deployment of a node fails, it's a transition made by MAAS itself.

So, from 'DEPLOYING', a node can be either manually released by a user, marked 'FAILED_DEPLOYMENT' by MAAS (if the deployment fails) or marked 'DEPLOYED' by MAAS (if the deployment succeeds).