[r=jtv,rvb][bug=1374388][author=jtv] Work around an “impedance mismatch” between Django's BooleanField in forms, and our use of forms for both UI and API.
The Node.disable_ipv4 checkbox could be enabled in the UI, but never disabled.  In the API it could go both ways.  The reason is that Django never submits a False value for an unchecked checkbox; it just doesn't include the value in the form submission.  Our forms are rigged to interpret missing fields as ones that should stay unchanged.  As a result, once disable_ipv4 was set to True, the UI simply ignored attempts to set it back to False.

The workaround adds an invisible input field to tell a UI-originated form submission from an API-originated one.  A UI form shows all fields, so a missing boolean field means “set to False.”  An API form must deal with callers omitting fields that were left out in order to keep them unchanged, so a missing boolean field does what it does now.  This is particularly important with fields that are added in later versions.