[r=rvb,allenap][bug=1375970,1375980][author=gmb] Refactor call sites of Node.objects.(start|stop)_nodes() so that they:  - Handle errors raised by those methods  - Ensure that model changes are committed before power actions are sent    (when starting nodes) and after power actions are sent (when    stopping nodes). This avoids race conditions when carrying out actions    in bulk.  - When errors occur in start_nodes(), the callsites will explicitly roll    the node back to a sane status (either its old status or a suitable    FAILED status).  - Re-raise those errors once they've finished with them  - Don't change the state of the Node if an error occurred whilst starting or stopping it.
This is a further fix for bug 1375980.

I've also removed the check for len(nodes_(started|stopped) > 0 in those methods that used it. This check made no sense; there are power types which we just *can't* power off (ether_wake, for example). If the check was left in place, the node would never come out of whatever state it was in, and would be stuck.

I've added tests to check for the atomicity of the callsites, for their logging and re-raising of exceptions, too.