[r=gmb,rvb,jtv][bug=1379154][author=gmb] Move the checking for boot resources out of start_up.py and into ExternalComponentsMiddleware.
This change came about because of Bug #1379154, wherein the link to the boot images page was a 404. This is probably (though I haven't verified the root cause of the bug) because when the check for boot resources ran in start_up.py, reverse() didn't know about any prefix with which to prepend the URL. Moving the check into the middleware continues an existing pattern of doing checks there, and also fixes the bug.

I've also removed the post-save triggers from the BootResource model, since they were no longer needed.