[r=kiko,allenap][bug=1379401][author=gmb] If we encounter an IntegrityError in EventManager.register_event_type_and_event(), catch the error and try to fetch the EventType again.
An IntegrityError will be raised when another thread has created the EventType before we called create(), but because it happened in a separate thread (and thus a separate transaction) Django doesn't know about it and doesn't create a ValidationError like you'd expect. Instead the error comes from the DB.

Our solution is not elegant, but should work for the majority of cases (and if we get a case where the EventType exists in the DB but doesn't get returned by get() then we've got bigger and weirder problems).

Another solution for this would have been to lock the table whilst we inserted the EventType row, but that just seems
 a) Like overkill
 and
 b) Not easy to do in Django-land.

We could shave yaks on this all day, but I suspect that this patch will fix the bug just fine.