[r=allenap][bug=][author=blake-rouse] Add storage constraint for acquire node selection.
Allows the acquire API call to specify a storage parameter to filter the available nodes. Parameter is structured as a comma-separated list of the minimum required size of disks in GB. Tags can also associated with a size, allowing better filtering based on the tags associated on the disks. Eg. 200,300(ssd,removable),400 - node must have at least 3 disks, first being at least 200gb in size with any tags, second at least 300gb in size with the tags ssd and removable, and third at least 400gb in size with any tags. The disks are ordered on the node by size providing the least greedy selection as possible.