[r=rvb][bug=][author=blake-rouse] Add triggers for when a tag is linked and unlinked from a node, and when a tag is updated.
This sends the node_update and device_update message when a tag is linked, unlinked, or updated. If linked or unlinked it will check if its installable. Installable sends a node_update message and non-installable sends a device_update message. When a tag is updated all nodes that are linked will get a notification message, so they get the new tag name.