[r=mpontillo,rvb][bug=][author=mpontillo] Automatically add hostnames when probe-and-enlist runs for vSphere.
This commit will use the name configured in vSphere for the hostname of the VM. In order to allow this, it adds an optional 'hostname' parameter to create_node(), and (if provided) passes it into the node creation form.

Also updated the create_node() test to also pass in that parameter.

MAAS will automatically add the DNS suffix, as applicable.

Note that this commit DOES NOT provide a way to turn this functionality off. If a vSphere VM is named with an invalid hostname, the node will fail to be added. We should determine if a use case exists for NOT using the VM names as the hostnames.