[r=blake-rouse][bug=][author=allenap] New function, callOutToThread(), which calls out to a given function in a thread.
This is similar to callOut() and is also meant for use in a callback chain. In addition, callOut's behaviour has changed to permit call-outs to asynchronous functions, and its usage has also changed: it no longer returns a closure.

This code has been extracted from work adapting MAAS to work with post-commit hooks with increased transaction isolation.