[r=andreserl][bug=][author=allenap] Suppress CancelledError just before cancelling a deferred hook.
Previously an errback was added just after adding the hook. However, this meant that callbacks added to the hook after it had been added to the DeferredHook instance could be called normally even though the hook had been cancelled. Put another way, the suppression errback must be at the end of the callbacks list.