[r=blake-rouse][bug=][author=allenap] Extract the way transition monitors are managed into a new class TransitionMonitor.
Previously, code running within the reactor that wanted to start or stop a transition monitor would have to call into a transactional context in a thread, which would then call back out to the reactor. TransitionMonitor makes it easy to define a monitor while in a transaction context that can be later used by code outside of a transaction context or in the reactor without having to switch context.