[r=rvb][bug=][author=allenap] Allow disconnection of field changes for testing.
Allowing tests to temporarily disable field changes makes testing post-commit code easier, because it avoids adding more boilerplate to neuter indirectly created post-commit tasks that are not the subject of the test.

This change also highlighted a test isolation problem, where the model change callbacks registered in one test were still present in other tests. Fortunately the problem was only in the tests; the code under test was fine.