[r=andreserl][bug=][author=blake-rouse] Add the set_active method to the Handler base class.
This allows the JS client to tell the region what object is the active object. That information will remain persistent as long as the websocket connection exists. This helps to make sure that full amount of information is not passed to the websocket. Only the active object will get all data. The makes the amount of data that needs to transfer over the websocket smaller, and makes the web application use less memory on the client side.