[r=rvb][bug=][author=allenap] In PostgresListener, only process notify messages from the database once.
Previously, when another message arrived before the previous batch had been processed, the previous batch could be processed a second time. This could cascade to multiple levels.

In addition, PostgresListener was not removing itself as a reader from the reactor when the connection failed. It's not clear that this was causing problems, but it makes sense to remove a broken connection from the reactor's purview.