[r=rvb][bug=][author=blake-rouse] Add EventHandler to the websocket and create triggers for the events table.
List events for a node with a maximum number of days. This is to limit the amount of events that are rendered on the client side, because at the moment we do not support server side searching. Adds triggers to the events table, sending signals for events and nodes. Also includes a fix on how the TimestampedModelHandler returns the created and updated fields.