[r=rvb][bug=1447208][author=blake-rouse] On update_power_state defer the releasing of static ip addresses on nodes.
This is done to limit the length of a thread that is running when a release action is performed. The action will be called later by the reactor with defered to a thread. This helps the release action to complete quicker and not blocking the thread that is currently running update_power_state.