[r=allenap,ubuntudotcom1][bug=1447739][author=blake-rouse] Send the correct notification messages to the websocket client based on the information the client has loaded.
This fixes issues where a user previously did not have permissions to view a node and now does. Instead of an update message being sent to the user a create message should have been sent. Same thing goes for losing permission to view a node, a delete action instead of update action should be sent to the client.