[r=rvb][bug=][author=blake-rouse] Remove the path field from BlockDevice and add it as a property.
This is done because curtin will make sure that every time a node is deployed or rebooted that the disks will always show up at this path. This is the best way we can set specific device names because udev will not allow you to set the name of the device at the "/dev" level.