[r=blake-rouse][bug=][author=allenap] Excise the broken two_dir_levels_up() function.
What two_dir_levels_up() was meant to do it get the parent directory of the given path, i.e. *one* level up. What it actually did is choose either the parent or grandparent directory, depending on the presence of a trailing slash in the given path.

The result of this was that downloaded images were going into /var/log/maas instead of /var/log/maas/boot-resources when the configured TFTP root directory did not have a trailing slash.

This has been fixed using Twisted's FilePath class, which makes the (fairly simple) path manipulations self-explanatory and predictable.