[r=blake-rouse,mpontillo,rbanffy][bug=][author=blake-rouse] Add bios_boot_method field to Node model and send the BIOS boot method from the TFTP service to the region when a node boots.
This stores the booting method of the node. There is no hard enum on the different types as those can change on the cluster and don't want to raise a database error if the region is not updated with that new value. When a node boots and requests its configuration we will record in the region its BIOS booting method. This method will be used for the storage layouts and to require a user to use the correct partition table type for their boot disk.