[r=andreserl][bug=1486926][author=blake-rouse] Fix create_physical_interfaces migration helper to set the created and updated.
This is needed because South creates the model object based on the current migration not on the code in src/maasserver/models. This prevents the code from getting any conveniences from the model, aka. setting created and updated for you. This was not caught in the tests because the tests need to pass in a model and the model that is passed in has this convenience.