[r=blake-rouse][bug=1445945][author=trapnine] Relevant node operations add a 'user request' entry into the event log.
An optional comment field is plumbed through the Node model methods and tests but aren't piped through the API in this branch (events-2 contains that change, MP pending).

User requests are displayed inline, hopefully before the events happening to the nodes. This allows one to see who (if anyone) triggered the events.