[r=mpontillo][bug=1517097][author=mpontillo] Make use of the "inner" dictionaries within LabeledConstraintMap objects passed into the specifier querying code. (that is, use the dictionaries for each label directly.)
Previously, the specifier querying code only recognized strings or lists.

With this commit, we use recursion within the specifier querying code in order to take the groups of constraints that the LabeledConstraintMap has organized, and ensure that each group of constraints is ANDed together. (in other words, we defer conversion to a list until we know what the grouping should be, since a LabeledConstraintMap implies a possibly-parenthetical query.)