[r=mpontillo][bug=1548601,1554566][author=blake-rouse] Use the new GetBootConfig RPC method in the TFTP service. Mark the node failed when the boot image on that rack controller cannot be found.
This work also made the code path for getting boot information about a node much simpler and faster. Previously the flow was TFTP -> HTTP -> RPC (back to rack that made the HTTP call). Now its just TFTP -> RPC and the rack controller figures out the boot images locally, telling the region to mark the node failed if it doesn't have that boot image.