[r=mpontillo][bug=][author=mpontillo] Get actions working for the forms on the networks listing page.
Drive-by fix for incorrect exclusion of spaces that contain no subnets from the space order-by on the network overview page.

Drive-by enhancement to the websocket dehydrate function so that it populates objects based on _id if a Model object isn't found.

Drive-by refactor/move of error message parsing code; it now lives in the ManagerHelperService.

Drive-by fix for VID validation, which was incorrect. (VID 4095 is reserved and should not have been allowed.)