[r=andreserl,ltrager][bug=1561816][author=blake-rouse] Ensure the same order of interface processing for the update_interfaces method for a rack controller.
The order of processing has to be the same no matter which regiond process is doing the process. This is because multiple can be doing the same rack controller at the same time. This will cause a deadlock because the first regiond process will grab a lock for its first interface and then the second regiond process will grab the lock on its first interface (which is really the second interface on the first regiond process), and a deadlock is created.