[r=blake-rouse,ltrager][bug=][author=newell-jensen] This branch updates the IPMI wait_time retries so that we don't keep getting "BMC Busy" messages from retrying too quickly. The last error is the one that ultimately gets registered to the event log so it is important that we get the actual reason for the original error.
Example of the issue from the logs:

/usr/sbin/ipmi-config: username invalid
ipmi_ctx_open_outofband_2_0: BMC busy
ipmi_ctx_open_outofband_2_0: BMC busy
ipmi_ctx_open_outofband_2_0: BMC busy

This would ultimately show "BMC busy" in the event log when the actual error was actually username invalid.  With this change (larger wait times in between retries) we get the actual error in the event log.

This branch also cleans up the error messages that are ultimately sent to the event log.

This was tested in the NEC lab and I saw really good performance with the wait_time used.