[r=lamont,mpontillo][bug=1571645][author=allenap] Drive DNS publication from the DNSPublication model.
Previously a sequence was being used directly for the zone serial. However, sequences are explicitly not transactional: in an HA environment, a zone published on one region host would not necessarily have the same serial as the same zone published on another host. Using a regular table ensures that all region processes use the same serial.