[r=mpontillo][bug=1571670,1572076][author=blake-rouse] Add the maasObjForm directive and use it to allow editing fabrics, vlans, spaces, and subnets.
The maasObjForm directive gives the ability to modify any object that is provided over the websocket. Provide the directive with the object you want to edit and the manager for that object. Define the fields in you want to be able to edit and profit! This follows the design that losing focus on the field causes it to save.