[r=blake-rouse][bug=][author=mpontillo] Add logic to detect which subnet a discovery occurred on, if one is present in MAAS.
Drive-by fix for a bug where the incorrect 'vid' column was being exposed in the discovery view.

This branch is necessary if we want to display the discovered devices on the subnet details page (or similar views). Since ARP observations can occur independently of a subnet known to MAAS, we make a good guess: if the observer's interface has matching subnet(s) configured in MAAS, we assume the observation was the longest-prefix-length subnet that matches the observed IP address found on the VLAN.