[r=blake-rouse][bug=][author=ltrager] Add support for hwe-rolling kernels
hwe-rolling kernels are metapackages which depend on the latest kernel available. Instead of duplicating the kernels in the stream lp:maas-images will determine if an already published kernel is a dependency of a rolling kernel and set the 'rolling' field to True. This MP stores that field in the BootResource table. When listing supported kernels the region appends the hwe-rolling kernels if one of the supported kernels has the rolling field set true. The rack now creates hard links for hwe-rolling kernels if the 'rolling' field is set true. This allows grub to show its booting a rolling kernel even though it didn't need to download a second set of files.