[r=blake-rouse][bug=][author=ltrager] Create Script, ScriptResult, and ScriptSet models.
The Script model will be used to store the default scripts as well as user created scripts in the MAAS database. These scripts will be run during commissioning or testing. The results of a script run will be stored in a ScriptResult which will replace the NodeResult model. Scripts run together will be grouped by a ScriptSet which allows users to easily view historical script results by run.