[r=blake-rouse][bug=][author=allenap] Run database tests without needing Django's test cases, mostly.
Previously Django's test cases did set-up and tear-down of the test database. In migrations mode this was teeth-grindingly slow; in legacy mode it was much quicker, but the database was not the same as would have been produced by migrations. Now a template database is constructed using migrations (with some cheats to make it much quicker) which is then used to prepare the database that tests use. In other words it combines the correctness of migrations with the speed of Django's legacy approach.