[r=ltrager][bug=][author=ltrager] Update the node-result API and websocket to work with the Script models.
This updates the node-result API and websocket to work with the Script models in a backwards compatible way. As discussed the node field only outputs the system_id of the node instead of the entire model in the API. The id between stdout and stderr is now the same. This is because we now store both in the same model.