[r=allenap][bug=][author=ltrager] Add MAASScript endpoint to the metadata service.
Adds a new endpoint at latest/maas-scripts which on GET returns an XZ compressed tar containing any commissioning or testing scripts which should be run by the requested node. An index.json file is also included which describes the suggested order the script should be run and what script_result_id the result should be uploaded with.