[r=newell-jensen][bug=][author=blake-rouse] Add ComposeMachine RPC call.
Allows the region controller to ask a rack controller to compose a machine that is greater than the `RequestedMachine` object. The pod driver should return the discovered newly composed machine and the updated pod hints. The ComposeMachine RPC call is very defensive in the results of the pod driver and does its best to ensure that the driver is returning the expected results.