[r=blake-rouse][bug=1626654][author=blake-rouse] Backport r5728 + r5737: Reap dead sibling regiond workers. Import message of degraded rackd.
This fixes users seeing bad reports of bad connections when dead regiond workers are stuck in the database for a longer period of time. As long as one regiond process is running on that region controller it will ensure that its other siblings are not in the database unless they are actually running.