[r=allenap][bug=][author=ltrager] Add test script reaper.
When testing begins a status_expires time of 20 minutes will be set. This gives the node 20 minutes to boot and begin testing, otherwise a failure is assumed. Once testing begins the status_expires is removed and the status_monitor checks for two things. First that the node is still up and pinging the region. Second, if a running script has a timeout it hasn't run past it. If either condition fails all pending and running tests are set to timed out status, the node is set to failed testing, and its shutdown unless SSH is enabled.