[r=mpontillo][bug=][author=blake-rouse] Snap up all of MAAS into one snap.
This creates the entire MAAS snap. Along with controls to modify the operation of the snap and check the status of the snap. Since snappy doesn't allow use to dynamically change the daemons that should be running we use supervisord to run all of the daemons. We change the daemons on the fly based on the mode of the snap and whether DHCP is enabled or not.