[r=mpontillo][bug=1377335][author=blake-rouse] Drastically reduce the number of queries required for a machines read.
Previously it was 700+ queries for 10 machines, now its <100 for 10 machines. All data is now prefetched when possible. Piston is monkey patched to fix an issue where it didn't use prefetched data. LogSQL fixture was used to identify the issues for this branch, it is included as its useful for doing this kind of thing.