[r=blake-rouse][bug=1686757][author=danilo] Support rack controllers using fully-qualified domain names as their hostnames
When a rack controller has a hostname set including a domain name (in the dotted form, eg. foo.example.com), registering this rack controller to existing MAAS regiond should keep the FQDN for the controller (if a domain does not exist already, it is added as non-authoritative domain name in MAAS).