[r=blake-rouse][bug=1686169][author=ltrager] Filter out unsupported products from BootSourceCache and send MAAS user_agent.
In the previous attempt the filter was applied whenever RepoDumper was used. Both the region and the rack use RepoDumper to list what images are available in a stream. The stream the region produces for the rack is similar, but not identical to the one at images.maas.io. These differences cause all images to be filtered out. Validation now only occurs when the region reads images.maas.io. The rack imports all images from the region and does not do any filtering.