[r=andreserl][bug=1690144][author=blake-rouse] Handle block devices in a virsh domain that is missing its backing.
When a virsh domain has a storage device thats backing file no longer exists or cannot be accessed no capacity is given for that storage device. Because of this virsh will also refuse to start the domain. MAAS does not add this domain because it cannot be used at all.