[r=blake-rouse][bug=][author=mpontillo] Make .idea directory creation consistent with .run directory.
Previously, there was a .idea directory checked into source control. However, the .idea directory is in the .bzrignore (and .gitignore) files. So it doesn't work quite right, because if you are actually using a code editor that supports IDEA project files (such as PyCharm), every time you open the project, those files will be updated, and the fact that .idea is ignored doesn't matter; those changes will still be present in your `bzr status`, and they're very easy to accidentally commit.

The solution is to make the .idea directory consistent with other skeleton runtime directories in MAAS (namely the ".run" directory).

This means developers wishing to use the IDEA project files will need to `make pycharm` (or `make .idea`) before launching an IDEA-based code editor.