[r=ltrager][bug=][author=mpontillo] Support for Fernet encryption using the MAAS secret as the key.
Fernet makes uses of AES-CBC and SHA256 to generate encrypted, authenticated, timestamped tokens. The tokens are in base64 format, and thus suitable out-of-the-box for usage inside a YAML, JSON, or BSON string for transport or storage. The full specification is available here:

https://github.com/fernet/spec/blob/master/Spec.md

The base64 token could be decoded if a more compact byte string is desired.

This initial implementation ensures the confidentiality of the MAAS key by using a high iteration count (100,000 iterations). It does not salt the key, since Fernet is intended to be used here to communicate on the network, and each peer may not necessarily have a secure way to share a salt.