[r=ltrager,blake-rouse][bug=][author=mpontillo] Add a helper script to join or leave the MAAS multicast groups.
MAAS will need to execute this on each monitored interface before listening for beacons, so that the NIC knows to forward beacon traffic up the OS (rather than dropping it before even copying it from the hardware buffer).

This approach is necessary in order to receive beacons on interfaces which may not have an IP address configured, and thus will not respond positively to socket-layer multicast join operations.