Backport from master: Run badblocks using the physical blocksize of the disk.
By default badblocks always runs using a blocksize of 1024 and uses a 32bit int for the number of blocks it can test. For large disks the number of blocks exceeds the 32bit int limit and causes badblocks to fail. Many large uses use a larger physical blocksize which avoids hitting the 32bit int limit which helps avoid hitting LP:1695229.