Backport d39a0e21fd14125a6cceb45cd692d34051f4a474 from master: Fix machinesTable directive so that it works with vs-repeat.
The machines, filteredMachines, and osinfo where placed on the scope directly. Once vs-repeat was added it changed the scope inside the directive to be the scope of vs-repeat and not the scope of machines-table. Moving those variables to the table variable on the machines-table scope fixes the issue as the table object will be copied by reference into the nested vs-repeat scope.