Add additional metadata fields to the Script and ScriptResult models to support results and parameters.
The Script model now allows scripts to define what hardware the script is for, if the script can run in parallel, what results the script will produce, what parameters are accepted by the script, and what packages need to be installed before a script is run.

The ScriptResult model now stores what parameters should be passed to the
script.