Update Script form to accept new fields and process embedded YAML.
The Script form can now accept and validate the hardware_type, parallel, and
packages fields. An embedded YAML configuration file may be embedded into the
script. The embedded YAML may set all script form fields in addition to the
results and paramater fields which may only be set in the YAML.