Use the FormScript when load script default and uploaded from the UI scripts.
Modified the builtin script loader to use the ScriptFrom. Builtin scripts now define their metadata in YAML embedded within the script. This includes planned support for results and paramaters which the scripts have not yet been modified for.

The CommissioningScriptForm now validates the script with the ScriptForm. Any included metadata will be parsed and validated. Commissioning scripts uploaded with an existing filename will now be considered updated.