Accept and validate results from nodes.
During commissioning and testing nodes may now send a result YAML file. The
result YAML may define the script result(passed, failed, or degraded) which
overrides using the exit_status to define the script_status. The result YAML
may also include results which is a dictionary of results from the script
which will be displayed over the API and UI. The result YAML file is
validated and an invalid YAML will be logged but still stored or later
viewing.

Nodes may also signal when installation begins. If a node signals failure
when the script_status is INSTALLING the status will be set to
FAILED_INSTALLING.