Backport 404d1299a735e57670609f27005f00cad7bc6c28 from master.
Fixes LP: #1711700 - Improve DNS reloading so its happens only when required.

This refactors all the the database triggers that alert all the active regions that DNS needs to be reloaded. The refactor ensures that DNS actually needs to be reloaded by checking to ensure that data was adjusted that requires a reload.

Trigger also now provide a proper message on what caused it to request the reload to be perform. The reload code now looks at the reasons and prints a log message each time it actually performs a reload for the reason it was performed.