Regenerate the current_testing_script_set when commissioning completes.
When commissioning finishes and there are testing scripts to run
maas_run_remote_scripts sends the TESTING signal. This changes node.status to
TESTING. Commissioning may have changed the storage devices available. MAAS now checks if any testing script to be run after commissioning uses the
storage parameter. If so the ScriptResult is deleted and a new one is created
which uses the default(test all storage devices).

maas_run_remote_scripts checks if any test scripts use the storage parameter
after the TESTING signal is sent. If so it redownloads the script tarball
which will contain the updated parameters.