Backport dd962958f73efb95216044f93de9aea3aace01e8 from master.
LP: #1507712 - cli: maas logout causes KeyError for other profiles

There is a race where when profiles are being deleted, if this happens
between the __iter__ and __getitem__ (of a different invocation of the
cli) which are two different SQL queries, there is an inconsistent view.

So instead use a cache with a consistent view of the data. This
guarantees correct -- if stale behavior in the cli.

Note that this is acceptable because when running two commands at the
same time ordering is unknown and the cache only lives for the lifetime
of a single command.