Fixes LP #1721548 - Failure on controller refresh seem to be causing version to not get updated.
Add maas_version to RefreshRackControllerInfo RPC response so the version will be sent over the wire to the region. Before the version was being calculated but it would not be sent over the wire. maas_version field is set to optional so older rack controllers that connect to the region do not error because that field will be missing.