LP: #1721824, #1721823 - Add health status
The UI now shows the health status of the node when the spinner isn't being
shown.  The health status is the overall status of all commissioning and
testing scripts.  Like the health status icons for CPU, memory, and storage
the health status icon will only display if there is a problem.

The commissioning, testing, CPU, memory, storage, and overall health are now
reported on the API. Like the websocket ScriptResults are cached so they are
only loaded once.