LP: #1727568, #1727575 - Create migration to show old storage tests on all devices.
MAAS 2.2 did not support assoicated a ScriptResult with a hardware_type or
physical_blockdevice. When the UI was updated in 2.3 it was decided that
these old results should be assoicated with all known PhysicalBlockDevices.
This was implemented in Javascript but has been causing issues with the
triggers as some ScriptResults are associated with a PhyicalBlockDevice while
others aren't.

Old ScriptResults are now assoicated with the node's known
PhysicalBlockDevices in a migration so the UI gets a similar data set.