LP: #1727547 -- Modifies the node model signals so that FAILED_TESTING is no longer blacklisted from updating the node's previous status.
This is needed so that we can actually go back to a FAILED_TESTING state when aborting tests.  Additionally, FAILED_TESTING is added to the NODE_TESTING_RESET_READY_TRANSITIONS so that in the event that the previous node status was FAILED_TESTING, that upon successful completion of running tests, it will go to the READY state.