LP: #1730799 - Make sure start time is always set.
When a script runs very quickly MAAS may receive the POST telling it the
script has started after the POST telling MAAS the result. If this happens
set the start time to the end time.

ScriptResults which do not have their start or end time set will now be
ignored when calculating a Script's estimated runtime.