Fixes LP #1748570 - Regiond caches the power and NOS types from the rackd controller per connection.
This reduces the load on the regiond and rackd because the regiond doesn't need to keep asking the rackd for the power type and NOS information. The missing packages is removed from the power types description so the cache is always consistent and missing packages will be reactive now (meaning when a power check is performed) instead of proactive.