Modify rescue mode to run commissioning or test scripts if available.
The script runner is now executed when transistioning into rescue
mode. Any commissioning or test script available will be run. Because
rescue mode allows the user to SSH into the ephemeral environment MAAS
does not check for loss of heart beat or remote checking if the script
has run past its runtime.

This does not implement the necessary API or UI changes to allow
users to select scripts to run in rescue mode.