Reduce triggers to only update the node websocket when viewable data changes
Instead of firing a trigger on any field change on maasserver_node only
fields which are viewable in the UI now cause the trigger to fire. This
uncovered a number of tests which were passing due to the previous catch
all maasserver_node trigger. The EVENT_NODE_NOTIFY trigger has it caused
events to update the node on the websocket. MAAS already has a handler
for events.