Don't pull in system libraries from the host into the snap.
When building the snap, snapcraft will automatically pull in libraries
from the system if it things they are needed. This means that the host
needs to be setup properly in order to build the snap. Better to have
the build fail so that you know that you have to include the library in
snapcraft.yaml.