Fixes LP: #1754493 - Call import boot images again if trigger for import when an import is already running.
This fixes an issue where the rack could still be performing an update with old data when the region triggers the rack to import again. This allows up to 1 import to run after the current import is finished. This will ensure that the rack controller will get always match the regiond images.