LP: #1759284 - Change the interface name gathering during commissioning
The interface names can now be larger than 8 characters before Bionic. Bionic forward these can be much larger than that. This caused issues where ifconfig -s would not output the whole interface name. As such, change the mechanism in which how we discover the interfaces for the DHCP commissioning script.