Signal skipped when block devices commissioning script is run in a container.
The block devices commissioning script runs on both controllers and machines.
It can not be run in a container or in a snap. When this is detected the script exists early returning no information. 720fa5c allows scripts to
return skipped when unable to run, this case was missed.

As this case can only happen on controllers support has been added for
controllers to signal results back.