LP:1701352 - Set default exit status to 1 when failure.
When Curtin sends the metadata server the installation result does not
include an exit status. It indicates pass or failure with SUCCESS or
FAILURE as the result. When storing the result if no exit status was given
MAAS assumed the script passed and set the exit status to 0. This caused
the ScriptResult for the installation log in both the UI and API to always
be passed even when Curtin failed and the node is marked failed deployment.
Now when Curtin sends failure 1 is set causing the ScriptResult for the
installation log to be set to FAILED. Commissioning and testing results use
a different API end point and are uneffected by this change.