Fix the get_pod_storage_pools method.
It called self.list_pools() with arguments, but list_pools() doesn't
accept any arguments.

Tests were passing, since list_pools() was mocked in the tests. I added
a new fake, which can be used to mock things at a lower level, so that
more of the functionality can be tested.