LP: #1762461 - Return list of operating systems from database instead of RPC
Previously whenever MAAS would list images or validate an image name it would
request for the list of images from all rack controllers. Each rack would run
the list of images it has an the title of the image according to an algorithm
in the osystem driver. Once the set of images on all rack controllers was
gathered MAAS would lookup titles for custom images in the database.

The RPC calls to all rack controllers caused the UI to block on some pages
as they require the GeneralManager to be fully loaded. The only reason the
images were requested from all rack controllers is to validate that an image
being deployed is available. Now all images names are retrieved from the
database. Before commissioning, testing, rescue mode, disk erasing, or
deployment MAAS validates that the request image is available on all
rack controllers. This eliminates RPC calls from the UI without increasing
the amount of database calls.