Remove StaticDriverTypesFixture again.
The StaticDriverTypesFixture was originall removed in 70f0c5b however it was
causing intermittent failures. These failures were due to get_all_power_types
returning a dictionary from the PowerDriverRegistry. As machines were added
this dictionary was modified and kept in memory allowing parameters to
leak between machine creates. This isn't happen over RPC because a new
dictionary was created on each RPC call. Now a deepcopy is used.