LP: #1707216 -- Match on pre-existing IPMI machine IP address while enlisting.
This fixes the issue where MAAS would create an additional machine if it didn't recognize the PXE MAC address even though a valid set of IPMI BMC credentials were supplied when creating the machine.  We now will return the pre-existing IPMI machine instead of creating a new machine when there is a matching IPMI IP address.