Make sure bond options are loaded.
2e3f1523 modified the UI to only load items from the GeneralManager when
needed. The node_details_networking controller is loaded only when on the
interfaces page. This is after the GeneralManager in the parent scope has
been loaded. Because the GeneralManager didn't know bond_options are needed
they were never loaded. Now make sure they are loaded before finish loading
the controller.