LP: #1782025 - Discover known machines during enlistment by MAC.
When a new machine is added with a MAC address that is not the boot MAC the
machine goes into enlistment instead of commissioning. When POSTing to create
a new machine MAAS will now search for existing NEW or COMMISSIONING nodes
that have a matching MAC address. If found that machine is returned instead
of attempting to create a new one. If the MAC address belongs to a Device or
Controller enlistment will still fail.