Generate the zone file for the MAAS internal domain.
Using the *connected* rack controllers a resource per subnet is created for the internal domain. Each resource (aka. subnet) has the IP addresses of all connected rack controllers on that subnet.

BIND was not happy with allowing '_' in the internal domain name. The branch switches to using 'maas-internal' as the internal domain name. The validation for the 'maas_internal_domain' setting has also been changed to use the default validation for creating/updating a  object in MAAS.