Use the maas-syslog service to output maas.log.
This uses the new maas-syslog service that MAAS configures and runs to output the maas.log. rackd and regiond now use the maas-syslog service as the logger for syslog. All region controllers will have the same maas.log file with all data from all the racks and regions in the same place.