LP: #1787621 - Retrieve undefined release title directly from OS registry.
MAAS tries to retrieve the resource title from the database. If it has not
been defined in the configured stream or by user during custom image upload
MAAS asks the OS driver to return a release. Each OS driver does this through
string manipulation and does not need on disk access to the OS images. As
such it is safe to call the OS driver directly from the region and not use an
RPC call to all rack controllers.