Add ability to clone interfaces from a source node to a destination node.
Uses similar logic to how cloning the storage layout works. Each clone occurs in a layer until the parent layer is cloned, only then can the child layer be cloned. This must take special care to be ensure that the MAC addresses are from the destination node and not MAC address from the source node.