LP: 1816452 - Loading network related page, such as 'interfaces' or 'add device' takes too long
The subnet.list websocket handler didn't scale with many observed IPs in
the database.

With the old code, subnet.list took 30 seconds when there were 40 000
observed IPs in the database. with the new code it takes 2.5 seconds.
There's a bug where we keep adding observered IP records in the database
with a null ip field. That was how we discovered the scaling problem in
the first place. But with the new code having those records in the
database doesn't affect performance.