Force lpxelinux.0 to use HTTP for requesting all files after the initial load of lpxelinux.0 from TFTP. All traffic after that initial load will use HTTP, including the extra binary files that lpxelinux.0 needs and the generated configurations.
This moves the bootloaders configuration from the top of dhcpd.conf into the subnet definitions, this mirrors what the dhcpd6 does. This is done because the path prefix includes the IP address of the rack controller, and it needs to be the IP of the rack controller for the subnet range it is providing.