Modify storage_layout_issues to allow ESXi to deploy with unsupported filesystems
This modifies Node.storage_layout_issues to allow VMware ESXi to be deployed
even when unsupported filesystems are defined. During deployment these
filesystems will be ignored. If no VMFS datastore is defined the default
datastore1 will be extended. Ubuntu, CentOS, and RHEL deployments are prevented from deploying when VMFS6 is defined as it is unsupported.