Fixes LP: #1818932 - Fix issues with loading machines and users over the websocket.
This fixes the machines and users handlers to be O(1) when using list. Batch sizing from the client side is also adjusted to provide further improvements to loading the listing based on each objects size.

This reduces the time to load 50 machines from 0.33 seconds(avg) to 0.030 seconds(avg) saving 30 seconds per list request.