LP: #1821220 - Prevent base ESXi partitions from being modified.
The VMware ESXi image is a DD image which contains a base set of 8 partitions
which may not be modified. This ensures onces the VMFS6 storage layout a
user cannot delete, format, or mount any of those partitions. To remove
them a different layout must be applied first.

VMware ESXi also skips partition 4 in its layout. MAAS was not properly
skipping this partition, this has been fixed.