Improve the check for checking whether a database is dirty.
The previous code tried to check, but it's was using an unreliable way,
and it ended up always marking the database as dirty.

Now we assume that the database is dirty, but allow tests to mark it as
clean, if it knows that the transaction was not committed.

MAASServerTestCase was changed to mark the database as clean, resulting
in the databases getting reused often and cuts the test suite run time
in half.