Allow Script parameters to be given by the user over the API.
When selecting commissioning or testing scripts over the API user may give
parameter arguments. These arguments will be validated and override the
builtin defaults. Parameters may be specified by the parameter name to apply
to all scripts. If a parameter is prepended by a script name that parameter
will only apply to the specific script overriding the default and any global
value.