Change config merge order to allow overrides
The  contains the user-supplied config from e.g.. Settings made in there are
expected to be passed to curtin as is. However, with other
(Maas-generated) configs being merged after it, values that the user set
explicitly may be overwritten by Maas.

Merging the  last ensures that the user-supplied values
override the Maas-generated ones.