Add the temp_expires_on field to the StaticIPAddress model.
Field is used for assignment of AUTO IP addresses. Field is set to a time in the future for IP address assignment, once the IP is checked and its free then the field is cleared. If the checking process where to crash a cleanup task is run to free any expired assignments based the desired time.