Modify maas-run-remote-scripts to apply custom networking when required.
When a script requires custom networking to be applied
maas-run-remote-scripts will now apply it. If any error is detected the
custom network configuration will be reverted and the script will be marked
as FAILED_APPLYING_NETCONF. Once the script finishes running the custom
network configuration will be reverted. Scripts which run in parallel will
now run in two groups, one which doesn't require custom networking and one
which does.