During commissioning generate a DHCP config if none has been created.
AUTOIP assignment happens as a post_commit() hook after a node
starts testing or deploying so MAAS can verify the IP address is
free. If it is requested during commissioning or in testing run
directly after commissioning no IP will be assigned when
networking configuration is reset. In this state generate a
configuration file with dhcp being run on the boot interface.
This is the same configuration run at boot so testing will be
done with the booted configuration.

Side fix - Make sure maas-run-remote-scripts detects dhcp configured
interfaces.