Allow dynamic hostname DNS records to map to multiple IPs.
Before this, every time MAAS updated a lease for an IP, it would delete
the dynamic hostname mapping and replace it with the currently
processing lease.

This mean that the dynamic hostname would alernate between IPs, if there
are more than on IP that has the same hostname. For example, if a
machine has two interfaces getting their IPs through MAAS DHCP.