Move DeferredAttribute loading prevention to a monkeypatch.
We currently prevent DeferredAttributes from being loaded from the
database in CleanSave.__getattribute__. However, that is very expensive,
since for each attribute get, we apply this check.

As an example, in a test that gets 30 nodes using the machine.list
websocket call, this patch took it down from 1.1 seconds to 0.6 seconds.