LP #1845459: Failure creating KVM pod in CI
Re-raise DatabaseErrors if form_pod.save() raises an exception.

The old code tried to mark the node as failed deployment, which failed
since the transaction was broken. Further more, the transaction wouldn't
be retried, which would have make the deploy succeed.

This is a stop-gap fix that improves the situation slightly. It's made
small so tha it can be backported to 2.6. The real fix is to refactor
the processing of the messages, so that marking the node as failed or
succeeded happens in a separate transaction.

(cherry picked from commit 6df0418010f49265549dac05233bc6cdfbc177e0)