LP: #1847046 - MAAS doesn't detect the right PXE interface
Extract the boot interface from the kernel cmdline.

This is more reliable than relying on the PXE boot MAC, since a lot of
cases we don't have it.

There's still code that tries to guess the PXE interface when the
machine boots with UUID, but I change it to keep the current one if it's
on the same VLAN.

(cherry picked from commit 5e371015372acd7f51901235188c74b2e1c25734)