Ensure that mimeparse is used from packaging and not from requirements.txt.
Previous commit added mimeparse to the requirements.txt which meant that during development mimeparse was being used from the virtual environment. Once installed MAAS was using the python3-mimeparse that doesn't have the MimeTypeParseException. This ensures that the development environment uses the mimeparse that will be used during installation and the snap.