LPP: #1867182 - When under load, MAAS snap keeps restarting services
Increase the timeout waiting for the the start/restart/status service
commands.

What happened before was that the start/restart command timed out, and
was re-executed. This served only to worsen the situation, since the
timeout stayed the same.

Given that the reason for having a timeout is to prevent a command from
hanging indefinitely, it's ok to have quite a high timeout. Having a
command hang completely is unlikely, and it's even more unlikely that a
retry will resolve the issue.