LP: #1873729 - Store and process host information from LXD.
When running commissioning, controller refresh, or getting information
from an LXD Pod include the host information in addition to the host
resources. This allows architecture and operating system information
to be collected in a consistent way across machines, controllers, and
Pods.

50-maas-01-commissioning now mimics the LXD API by providing host
information in the root of the JSON object while resources is a
subobject.