LP: #1878117 - Fix LXD Pod storage amount.
LXD composed VM's were being created without setting the disk's model or
serial. Commissioning detects a disk with a model and serial so MAAS
removes the originally created disk and creates a new one with a model and
serial. This removes the storage pool link which prevented MAAS from
determining the amount of storage used in the pool.

Previously MAAS calculated the total storage amount for an LXD Pod to be
the total available disk space. It is now calculated based on the amount
available in associated storage pools.