LP: #1876855 Add can_probe to the driver power classes
This boolean flag determines if the power driver is suitable for use with
add_chassis.

The http API has been updated to use this flag when validating the allowed
list of driver names and the websocket API now includes this flag when
returning the list of power drivers from the general.power_drivers call.

To support backwards compatibility a new alias field was added to the driver classes. This allows them to be referred to by other names. For example "powerkvm" for virsh and "seamicro15k" for "sm15k"