LP: #1888536 - named.conf.options.inside.maas reverts to default
run-named always generated a blank config for bind when it's started.

It's questionable why we need to do that on every start, but for now, we
don't generate a config if a config is alrady there.

This should be safe enough to backport, and then we can revisit this
later, to see if we can fix it in a better way.

(cherry picked from commit 192fd8b70d0524683a50f0510577bdb1dc89ec5a)