Bug:1883748. Handle PermissionError in atomic_write()
If for any reason the permissions of the original file cannot be cloned
and set in the temporary file the PermissionError will be raised, before
this patch the exception will be unhandled leaving the temporary file
behind, which for situations where the the region or rack controllers
will retry the generation of said file (e.g. maas-proxy.conf) will
produce hundreds of files in a few minutes.