Load all scripts into the database before creating the region in the database.
When a new MAAS is added one of the first things that happens is the new
region is added to the database as a node. Creating a controller node
object also creates a ScriptSet and related ScriptResults to store
commissioning scripts from the region which are run later in startup.
fe5acf1 moved all commissioning scripts into the database. Because
load_builtin_scripts() was being called after the region was created
the region was never able to be created in the database causing startup
to fail.