Accept ScriptResult status updates by name and accept runtime from client.
During enlistment the client receives commissioning Scripts before the
Node or ScriptResult exists. As such the client doesn't have a way to
pass ScriptResult.id once the Node is created. MAAS already accepted
ScriptResults by name, this allows status changes to be passed by name
as well.

Some Scripts will finish running on the client before the ScriptResult
or Node objects are created. In this case MAAS will miss the signal that
the Script has started. When the result is uploaded the client may include
a runtime, MAAS will calculate ScriptResult.started to be
ScriptResult.ended - runtime so the runtime is preserved.