Enable using maas-run-remote-scripts for enlistment.
MAAS will now send the same user_data for all ephemeral environments
except disk erasing. Enlistment is handled by maas-run-remote-scripts
and BMC configurations is expected to be done in commissioning scripts.

The skip_bmc_config column has been removed as commissioning Scripts won't
be added to the ScriptSet when commissioning is started.

Enlistment was previously using a seperate /enlist enpoint for anonymous
access. This is still available but the normal metadata endpoint can now
be accessed anonymously. This removes the need for URL mangling.