Correctly discover the NICs for LXD VMs.
The code for discovering which NICs a LXD VM is connected to was a bit buggy.
It passed the name of NIC inside the VM as the host NIC to which it's attached
to. Further more it didn't convert "bridged" to "bridge", which is the correct
attach type.

I also took the opportunity to git rid of some testtools usage.