Amend the logic for matching VM NICs with host NICs to account for SR-IOV.
I also had to change the way we generated the NIC devices when
composing, as well as how we discover them. Since SR-IOV VFs are passed
on to the VM as PCI devices, LXD doesn't know the MAC when the VM is
turned off. Furthermore, each time the VM starts, it might get another
VF, with another physical MAC.

We might want to deal better with the VF MACs in the future, but setting
the MAC address explicitly allows such VMs to be used in MAAS for now.