LP #1899486: Initial storage configuration doesn't take EFI into account.
There's code to update the boot method for a machine when it loads the
ephemeral environment. But when a machine is added with IPMI, we don't
have the MAC address, so we can't identify the machine.

I think we should solve this in a better way, but put in a simple
workaround for now.