LP #1900832: DCHP can fail to start due to a race condition
There's a delay between asking for listening to postgres sys_core message and when the listener is actually registered in postgres.

We now explicitly wait until the process is actually listening until we
move on with dealing with possible missed messages.

I'm not 100% sure this actually fixes the problem, but I have high hopes
that it will.