Change the default OS to 20.04 LTS, Focal, on new installations.
The libvirt-bin metapackage was deprecated in 18.04 and removed from
20.04. MAAS now installs libvirt-daemon-system and libvirt-clients
which were previously installed by libvirt-bin. qemu-efi is no longer
installed as MAAS does not compose UEFI VMs with libvirt.

The websocket has also been updated to allow KVM hosts to be deployed
on any Ubuntu version. The API already allowed this.