Retry IPMI power commands without K_g key on K_g key failure.
Some BMCs will claim they are IPMI 2.0 complainant, allow you to set an
IPMI K_g key, allow you to verify an IPMI K_g key is set, yet will deny
any IPMI requests that use an IPMI K_g key. The error given is the K_g
key is invalid. This can be very confusing to users as MAAS allows you
to set an K_g key leaving them to think the only fix is to reset the
machine in person. In this case simply not using an K_g key allows all
IPMI commands to work again.

When there is an IPMI failure which is due to an K_g key and an K_g key
is set MAAS will retry without the K_g key. This is logged as a event
so the user knows what is happening but can still use their machine.