LP #1903673: Deploying a machine sometimes takes a long time
The 'ping' command would simply hang sometimes and the RPC call to check
the IP would time out after 2 minutes.

The most likely culprit is that when spawning the ping process, twisted
is acting up in a weird way. I haven't been able to pin-point the
problem, though. The only way I could get it to work was to not use
reactor.spawnProcess.