replace many direct uses of deferToThread by deferring entire async functions
Instead of doing many back and forth via deferToThread for pylxd calls, defer whole functions to be run in a separate thread.
This simplifies a bit the code and makes it safer against calls to pylxd.Client which end up making http calls.
It should also reduce a bit the overhead of thread communication, although it might block the thread for a longer time doing work.