LP: #1902425
 - always create neighbour entries for IPs in use.
 - exclude already tested IPs when claiming IPs during deploy

Rework Interface.update_neighbour move check for neighbour discovery to call sites.
This also changes IP claim checks to always create a neighbour entry when the IP is found to be in use (even if the rack interface doesn't have neighbour discovery enabled).

This fixes a corner case for the linked bug which happens when the used IP is
on an interface that has neighbour discovery disabled, in which case the IP is
not recorded in MAAS, and can get picked up again.