Run the commissioning scripts in the rack network interface monitoring service.
This is the first top towards unifying the way commmisioning scripts are
run in both regiond and rackd. Another branch will refactor the base
network interfaces monitor service so that both regiond and rackd will
do the refresh there.

The reason it's done in the network monitoring service is that that we
want to add the hints to the commission output and move the interface
processing there. This will allow us to use the same code for both
controllers and LXD VM hosts.