LP: #1887558 - Condense LUNs when detected by commissioning.
Multipath devices are shown on Linux and detected by LXD as one block
device per path. There is still only one backing block device, just
multiple ways to get to it. If the block device is a multipath device the
device_path will contain what LUN it belongs to. Each LUN is unique to the
backing device. When this is detected condense all paths into one block
device for MAAS to model. On deployment Curtin will detect this is a
multipath device and configure it properly.