Annotate controller commissioning output with extra info.
This is that we can move the logic from Controller.update_interfaces()
to the commissioning hook.

I named it "network-hints" for now, but that will most likely be renamed
before the release. The plan is to change the data itself, so that we
don't send anything that can be derived from the normal commissioning
output.