Add set_boot_order as a base ability to power and pod drivers.
Power and Pod drivers may optionally implement the ability to change boot
order. If supported, as indicated by the can_set_boot_order variable, the
region will send an ordered list of network and storage devices
representing the expected boot order when booting or when a machine hits
the netboot_off endpoint. The list is the serialized object containing the
id and name storage devices also provide the id_path, model, and serial
while network devices contain the mac_address, vendor, and product.