Process the network interfaces only in the commissioning hook.
This gets rid of Controller.update_interfaces, so that we can be sure
nothing is calling it.

It also doesn't update interfaces in maasserver/startup.py, but rather wait for the network interfaces service to come up, nor does it update interfaces when the rack registers.

Now there's only one place where we update networking information.