Get rid of redundant tests in TestUpdateInterfaces.
There we scenarios setup, so that each test was being run 6 times. But the
scenarios didn't make sense, since there was nothing in the implemenation that
cared about the different scenarios, except for a few tests that cared about
beaconing.

For now, the functionality is mostly preserved. Instead of running each
scenario, I made it so that the tests random pick what to test with, since it
doesn't matter.

For beaconing, I duplicated a couple of tests, until I have time to clean up
the tests properly.