Call set_boot_order when starting or toggling netboot.
MAAS will now automatically call set_boot_order when starting a Node
or when toggling netboot with the metadata service API. This allows
Nodes with supported boot driver to be dynamically configured to
boot over the network or boot locally without having to network boot
first. set_boot_order is only called if the power driver supports
the call.