add service to track and report snap versions
This is the first in having rackcontrollers reporting information about snap versions.
Currently this is not yet available from snapd inside the snap, so this emulates collecting information by reading a file.

Versions are reported to the region which for now only logs the information (until needed model changes are in place).