Add a script to make it easier to modify a MAAS running as a snap.
It check the git status for src/maasserver, src/provisioningserver and
src/metadataserver. If any file is modified, it will create a bind mount in
/snap/maas/current/lib/python3.8/site-packages/ for that.

You should check out the commit of MAAS that the snap is based on. For
example, if the local installed snap is 2.9.2-9165-g.c3e7848d1, checkout
c3e7848d1. Otherwise, the file you change might depend on something that
isn't in the snap.

It uses 'git ls-files -m' to determine which files have been modified.