LP #1929086: LXD VM hosts can't be refreshed if VLANs interfaces aren't named .
This is a small fix that can be backported to 3.0.

We should remove the automatic generation of of VLAN interface name
completly, since the callsites should be responsible for giving the
right name. That's a bit too risky of a change for 3.0, though, since it
breaks a lot of tests.

(cherry picked from commit 38616374f7eec730d6e90fc5ec9b5fcb46695f8d)