LP #1931215: two IPs assigned to one interface
While the root cause most likely is our primary/secondary DHCP setup,
this works around the case when the machine gets two different IPs from
the DHCP requests.

Even if we see multiple IPs for commissioned nodes, we only create one
link to the subnet.

In the case where the machine is a controller, we still record both the
IPs.