Return all the commissioning scripts when the machine is deployed.
Before, if you got the commissioning scripts for a deployed machine from the
metadata server, you could get an empty list.

To support updating the hardware information for deployed nodes, we now return
the finished scripts when the machine is deployed.

I also added a has_finished attribute to the script metadata, so that we can
change our script runner to check that no scripts need running, instead of
checking that no scripts are available. That way way we can simplify the logic
and always return all the scripts.