Make automatically created nodes for controllers DEPLOYED.
I made use of the 'dynamic' flag on machines to make it a bit easier to see
that MAAS automatically created the machine. The 'dynamic' flag is currently
only used for VMs, but I think it makes sense here as well.

The reason for this change is to get rid of all the places where we check for
is_controller() when processing commissioning data and instead check for status
== DEPLOYED, so that we process things the same, no matter if it's a
controller, VM host, or any deployed machine.