Don't throw away name given to VLANInterface.
When a VLANInterface was created, the name got thrown away and a new one was
generated, unless the node was a controller or VM host.

When creating a VLANInterface, the name should already be known and trusted. I
moved the logic for generating the name to VLANInterfaceForm, to support
creating a vlan interface without giving a name in the API.