Record endpoint in 'maas-run-scripts register-machine'
When a machine was registered, the API tokens were written to a file.
But the metadata endpoint wasn't included, so you had to specify that
manually when running 'report-results'.

Now all that you have to do is to pass the yaml file.

We still have the issue where the commissioning script expects the
resource binary to be downloaded from the parent of the metadata server.
For now, you have to manually change the port to 5248 to make it work.