Add tests ensuring that pod.get_projects accepts credentials.
The pod.get_projects websocket method already can pass through the
certificate and private key.

This branch adds test to ensure that it works, and will continue
working. The mocking is moved to the RPC layer, so that this test will
start to fail if we change the RPC call to be more strict in the future.