Catch not found errors when generating commissioning data for vm hosts.
We first get the list of network interfaces, and then for each interface
we get the state of it. During that time, some interfaces may have
disappeared.

This especially happens when adding a vm host, since the VMs are
starting and shutting down, which causes tap intefaces to be removed.

It should be enough to catch NotFound errors, but due to a bug in LXD,
we have to catch LXDAPIErrors as well.