Another attempt at fixing dblocks intermittent failurs.
While the previous fix improved things, we're still seeing intermittent
failures now and then.

Now, we use a DB sequence to ensure that each test use unique objid.

But also, that change alone actually made it fail more consistently. So
in addition, I fixed get_locks() as well. It got all the locks across
all the databases in the cluster, and not only the current database.
That meant that it got locks from other test processes as well.