Don't gzip responses in Django.
This caused the image sync from region to rack to slow down considerable
and use up a lot of CPU resources.

On my machine, with rack and region on the same machine, it took around
15 seconds to sync one 450 Mb image, with regiond using 100% CPU most of
the time.

With this change, the sync takes only a couple of seconds, and the CPU
spike is gone.