Bug #1853047: VLAN with the specified VID already exists error when updating the fabric attribute
Don't create a new VLAN for an existing VLAN interface if we can't
autodetect what VLAN it should be.

If you move multiple VLANs via our UI or API, you can only move one at a
time. This gets you into a situation where an interface and its VLAN
interfaces may be on different fabrics.

If the controller refreshes it network information, it shouldn't try to
change things back to how it was before the move.