Remove unused code from Node.set_power_config().
There was code that checked for self.bmc, even though it couldn't really
be None, except for in unit tests.

A lot of tests passed in "" as the power type, expecting to get a node
without a BMC configured. But in reality, that created a BMC with power
type "", which is invalid.

(cherry picked from commit eb055f00437974e4a36ea2969002b2830f81af9f)

Add stats for power drivers.

Also added an attribute to BMC, so that we can track whether the BMC was
auto-detected in commissioning, or whether the user created it manually.

(cherry picked from commit ab360b09b24e750d106e07778574e72a8dfc99fc)