962449ecd backport: set the IP address, not the ID in exclude list
IDs were erroneously added to the list of IPs to exclude. This was not failing
before because the netaddr library was accepting "entry in network" with entry
being an integer (which gets silently converted to an IP).
Newer netaddr versions actually blow up on this.
The exclude list was intended to contain actual addresses, not database IDs.