Log details when updating ipmi power_type machine
In a scenario where the IP address of two IMPI machines over lap the
details of the latter silently override the former.

This logs the update when it takes place so that the user can make an
informed decision about whether this is correct of if they have a
miss configuration present.