LP #1955709: Metadata field may_reboot not working correctly in 20.04
When a commissioning script needs to reboot, it simply calls 'reboot'.
While that is happening, cloud-init fails and notifies MAAS about it,
and the script runner happily continues running the scripts, since it
doesn't know about the reboot.

The following happens:
  1) Script calls reboot
  2) cloud-init notifies MAAS that it failed
  3) Script runner marks the script as PASSED
  4) Machine reboots

In Ubuntu 18.04, it seems that this is what is happening. In 20.04 and
forward, the order of events are changed slightly, so when MAAS gets
the notification that cloud-init failed, the script is already marked as
PASSED.

I've changed the logic to account for this, but the real fix is to stop
scripts from directly calling 'reboot' and instead tell the script
runner that a reboot is necessary.