Make it more explicit what to measure in the performance tests
Before, we measured how long the whole test took to run. That includes
any local setup you do in the test, but also the setup of any fixtures
the test uses.

I also combined measuring the time and, optionally, do profiling
using cProfile. It's more logical that you want to profile the
thing that you are actually measuring.