Control the garbage collection for the performance tests.
Before, one test could create a lot of garbage that then got
collected while running another test. That made the second test
look slower than it actually was.

Now we collect the garbage before a test being run, and also
collect it after the test is being run, including the last collection
in the timing of the test.