Bug #2003310: Refresh scripts are not re-run if they pass, but fail to report the results to the region
While refreshing the hardware information for controllers,
the scripts might all pass, but when reporting the results
back to the metadata server, something went wrong.

The exception was swallowed and ignored, though, so the
networking service didn't noticed it and assumed that
all scripts had passed.

This changes things so that the signal exception is caught
in the networking service itself, so it's aware that something
went wrong.