fix: change database owner
MAAS requires $dbc_dbuser to be an owner of $dbc_dbname.
One of the reasons is to be able to create new schemas (CREATE SCHEMA).

This is not new and documentation mentions the following configuration step:
```
sudo -u postgres \
  createdb -O \
  "$MAAS_DBUSER" \
  "$MAAS_DBNAME"
```

Unfortunately dbconfig-common used by MAAS configures `postgres` user as the
owner of `maasdb`.

Starting from dbconfig-common 2.0.22 the owner of the database
will be PostgreSQL user ($dbc_dbuser) [0]. For us this is the desired behavior.

Because dbconfig-common 2.0.22 is not available in 22.04, we have to fix
it ourselves for new installations and handle this case during upgrades
via deb packages.

In situations when MAAS cannot fix database ownership automatically we
want to inform MAAS user via deprecation notice [3].

[0]: https://salsa.debian.org/debian/dbconfig-common/-/blob/master/debian/changelog#L21
[1]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1010923
[2]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1021791
[3]: https://maas.io/deprecations/MD4