Don't transition to FAILED_EXITING_RESCUE_MODE too early.
When exiting rescue mode, the node is either turned off or power cycled
in a post commit hook. Then, every 15 seconds, we check whether the
machine is on or off and updates the machine status.

If that check happens while a power cycle is happening, or before
the machine is actually turned off, the machine would go to
FAILED_EXITING_RESCUE_MODE incorrectly.

We now wait some time to elapse since the start of exiting rescue
mode before marking the node as failed.

This is a temporary fix until we write a workflow for this.