fix(redfish): handle Reset status in Redfish power driver.
HPE Gen11 (and above) BMC might return a "Reset" or an "Unknown" status. Such statuses are transitional, so the BMC will eventually return another "known" status.
For such reason, with this change the MAAS Redfish power driver is going to retry up to 7 times to get a "known" status. If the maximum number of retries is reached, an error is returned.

Resolves: #LP2079797