refactor(dhcp): remove ValidateDHCPvXConfig RPC call
These RPC calls were used for validating DHCP Snippets by running dhcpd on the Rack Controller.

Since we plan to replace the current RPC solution and DHCP Snippets is a less frequently used feature in MAAS, we've decided to remove the validation logic as part of our shift to an alternative approach.