feat(provisioningserver): Added power_reset method to power driver base class
In preparation for adding DPU support into MAAS, it was necessary to add a power reset method to the base class of the power driver to be able to properly manage the DPU's power state. DPUs cannot be hard reset - powered on, off, or cycled - without restarting the whole chassis they're plugged into, so instead they're soft reset (their OS rebooted). This base functionality is the first step in supporting this.