fix(redfish): Unset If-Match header before power action requests.
The Redfish power driver reuses the same context and headers for all requests in a power action. However, when setting PXE boot, the `If-Match` header is included, causing subsequent requests to use the same value. The entity changes in the meantime and some BMCs return a 412 (Precondition Failed) error. To prevent this, ensure the `If-Match` header is unset before making the request.

Resolves: LP#2099949