test: Added unit tests for power activities in maas-agent
These previously were not unit tested. This patch modifies the `exec` calls in `powerCommand` and `PowerService` to use interfaces, with the production ones set in `PowerService.configure` and mock ones being test in the accompanying unit tests. Both the `CommandContext` and `LookPath` methods had to be replaced with injected dependencies.