fix: set DeployMany wf execution time to infinite to handle failed deployments
The DeployManyWorkflow and its child workflows (DeployWorkflow) had the
same execution time, resulting in the former being terminated before the
latter. Since the logic to mark a node failed is inside the parent, when
it times out (and consequently times out all the child workflows), the
node won't be marked as failed.
To avoid this, we set the execution time to infinite (the default when
not specified) since we are sure that eventually the parent workflow
will terminate when all the child workflows terminate.

Resolves LP: #2103733

(cherry picked from commit 95f40f18c317f9d42b4f5687c72e24cc24b55047)