refactor: BootResource API upload
There were two different forms for uploading a boot resource:
- BootResourceForm
- BootResourceFormNoContent

The first one was accepting a `content` parameter which represented the
custom image being uploaded. But, due to our nginx config, the maximum
allowed size for the HTTP body is 10MB, which is too small for every
custom image we might want to upload.
The second one, instead, will only create the necessary entities in the
db. The user then will have to upload the image in chunks.

The following changes have been made:
- removed the content parameter from the API
- the BootResourceFormNoContent now becomes the BootResourceForm, as
  this is the only way to create a boot resource
- the API doc for file upload in chunks have been made visible in the
  API docs
- Tests have been adapted to the changes made
- The method `create_from_content` has been removed from the
  LocalBootResourceFile

BREAKING CHANGE: removed the `content` parameter from the boot resource
create endpoint.