feat: save the simplestreams manifest in the db
This commit introduces a new table `maasserver_imagemanifests` that will store
the simplestreams manifest in the db as json, one for each boot source.
The logic to retrieve one would then be: first check in the db, if not present
fetch the simplestreams mirror.

It does not implement a BaseRepository/BaseService as we don't need all that methods.

As a consequence of the addition of this service, a part of the methods in the
image sync service have been moved to the new image manifests service.