fix: fetch manifest schedule execution
In order to register schedules the Temporal namespace must be already
registered.

Also, the image sync workflow is handled by the twisted TimerService and
we have no guarantee that it will be executed after the fetch manifest
schedule (resulting in MAAS not synchronizing images on the first
initialization). To avoid this, execute the fetch manifest workflow
when initializing the timer service.