feat(bootresources): implement a db view for querying the status of a selection
This commit introduces a new db view to be able to query the status of all the selections. In detail what you can see is:
- status: merged status of the boot resources that composes the selection
- update_status: whether there is an update available from the stream for any of the boot resource inside the selection
- sync_percentage: percentage of file size synchronized
- selected: whether the selection is the one with highest priority

The id in the view is always the selection's id.

A repository and service for it has been implemented, together with the API v3 handler to query the status for a particular selection.