fix(bootresources): revert boot source selections v2 API breaking changes
Now both the new and legacy boot source selection model co-exist in the db.
The v2 models/forms/ws/api are the same as before.

To ensure the compatibility of both models, in v3 we have to do some extra steps when creating/updating/deleting new selections. This actions are all handled inside the
boot source selection service.

Also, to be sure that legacy selection and new selection are in the same state, we ensure
this every time the fetch-manifest workflow is executed.